requirements.txt
# requirements.txt
openai>=1.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pyyaml>=6.0.1
